added a new file for text field picker
